₦Ҿ₢⃝№щΐ₵ᴑᴎ⁸ Ver.8 Rev.E by VxD
NECR0N0MiC0N 8 can change the look and feel of mIRC. Customize nicklist, Treebar dialog boxes, popups and much more...
Is a software that allows you to manage the graphics of the mIRC client in many aspects. Apply styles, colors, images and transparencies to the pop-up dialog boxes and windows has never been so easy, 100% customizable.

This Script is evolving to enable any combination of style, color and aspects of graphic display effect to your dialog window with DCX.dll and other notes Dlls combined with each other. Is possible to update the script with update system for all versions or addons that will be introduced later.
Optimized for Windows 10, then activate Aero style glass or Glow to get the maximum effect in the dialog boxes.
The minimum requirements for the installation are:
 ◘ Windows 10
 ◘ .NET Framework versions 4.5